pyproject.toml
[tool.poetry]
name = "bridge"
version = "0.1.0"
description = ""
authors = ["Ярослав Андреев <1frost34eone@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
django-environ = "^0.11.2"
gunicorn = "^21.2.0"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}
pillow = "^10.1.0"
drf-extra-fields = "^3.7.0"
django-filter = "^23.3"
django-cors-headers = "^4.3.0"



[tool.poetry.group.dev.dependencies]
black = "^23.10.0"
ruff = "^0.1.1"
isort = "^5.12.0"
pre-commit = "^3.5.0"
factory-boy = "^3.3.0"
coverage = "^7.3.2"
pytest-django = "^4.5.2"
pytest-cov = "^4.1.0"
pytest-factoryboy = "^2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
